+++ lib/flip_layout.dart
@@ -1,3 +1,4 @@
 export 'src/layout_motion.dart';
 export 'src/animated_layout.dart';
 export 'src/motion_group.dart';
+export 'src/motion_config.dart';